requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Deep learning (PyTorch)
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=10.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Progress bars
tqdm>=4.65.0
